+++ api.planx.uk/modules/saveAndReturn/service/resumeApplication.test.ts
@@ -1,8 +1,11 @@
-import { LowCalSession, Team } from "./../types";
+import { LowCalSession, Team } from "../../../types";
 import supertest from "supertest";
-import app from "../server";
-import { queryMock } from "../tests/graphqlQueryMock";
-import { mockLowcalSession, mockTeam } from "../tests/mocks/saveAndReturnMocks";
+import app from "../../../server";
+import { queryMock } from "../../../tests/graphqlQueryMock";
+import {
+  mockLowcalSession,
+  mockTeam,
+} from "../../../tests/mocks/saveAndReturnMocks";
 import { buildContentFromSessions } from "./resumeApplication";
 import { PartialDeep } from "type-fest";
