+++ api.planx.uk/modules/saveAndReturn/service/validateSession.test.ts
@@ -1,7 +1,7 @@
 import supertest from "supertest";
 import omit from "lodash.omit";
-import app from "../server";
-import { queryMock } from "../tests/graphqlQueryMock";
+import app from "../../../server";
+import { queryMock } from "../../../tests/graphqlQueryMock";
 import {
   mockFlow,
   mockLowcalSession,
@@ -11,10 +11,10 @@ import {
   mockGetMostRecentPublishedFlow,
   stubInsertReconciliationRequests,
   stubUpdateLowcalSessionData,
-} from "../tests/mocks/saveAndReturnMocks";
-import type { Node, Flow, Breadcrumb } from "../types";
-import { userContext } from "../modules/auth/middleware";
-import { getJWT } from "../tests/mockJWT";
+} from "../../../tests/mocks/saveAndReturnMocks";
+import type { Node, Flow, Breadcrumb } from "../../../types";
+import { userContext } from "../../auth/middleware";
+import { getJWT } from "../../../tests/mockJWT";
 
 const validateSessionPath = "/validate-session";
 const getStoreMock = jest.spyOn(userContext, "getStore");
