--- api.planx.uk/saveAndReturn/index.ts
@@ -1,4 +1,0 @@
-import { resumeApplication } from "./resumeApplication";
-import { validateSession } from "./validateSession";
-
-export { resumeApplication, validateSession };
